ConfigurationResolver - move into Console subnamespace
